+++ packages/core/src/@types/session.ts
@@ -1,16 +1,10 @@
+import { EditableShape, ShapeToObject } from "./utility.ts"
 import type { TypedJWTPayload } from "@aura-stack/jose"
-import type { CookieStoreConfig, InternalLogger, JoseInstance, RouterGlobalContext } from "@/@types/config.ts"
+import type { UserIdentityType, UserShape } from "@/shared/identity.ts"
+import type { CookieStoreConfig, IdentityConfig, InternalLogger, JoseInstance, RouterGlobalContext } from "@/@types/config.ts"
 
-/**
- * Standardized user profile returned by OAuth providers after fetching user information
- * and mapping the response to this format by default or via the `profile` custom function.
- */
-export interface User extends Record<string, unknown> {
-    sub: string
-    name?: string | null
-    email?: string | null
-    image?: string | null
-}
+export type User = UserIdentityType
+export type { UserShape } from "@/shared/identity.ts"
 
 /**
  * Session data returned by the session endpoint.
@@ -216,11 +210,12 @@ export interface GetSessionReturn<DefaultUser extends User = User> {
     headers: Headers
 }
 
-export interface CreateSessionStrategyOptions<DefaultUser extends User = User> {
+export interface CreateSessionStrategyOptions<Identity extends EditableShape<UserShape>> {
     config?: SessionConfig
-    jose: JoseInstance<DefaultUser>
+    jose: JoseInstance<ShapeToObject<Identity> & User>
     cookies: () => CookieStoreConfig
     logger?: InternalLogger
+    identity: IdentityConfig
 }
 
 export interface JWTStrategyOptions<DefaultUser extends User = User> {
@@ -228,6 +223,7 @@ export interface JWTStrategyOptions<DefaultUser extends User = User> {
     jose: JoseInstance<DefaultUser>
     logger?: InternalLogger
     cookies: () => CookieStoreConfig
+    identity: IdentityConfig
 }
 
 export interface SignInOptions {