+++ packages/core/src/index.ts
@@ -3,6 +3,7 @@ export { createJoseInstance } from "@/jose.ts"
 export { createSyslogMessage } from "@/shared/logger.ts"
 export { builtInOAuthProviders } from "@/oauth/index.ts"
 export { createAuthClient, createClient } from "@/client/index.ts"
+export { createIdentity, UserIdentity, StrippedUserIdentity } from "@/shared/identity.ts"
 
 export type * from "@/client/index.ts"
 export type {
@@ -40,4 +41,9 @@ export type {
     StatelessStrategyConfig,
     SessionConfig,
     SessionStrategy,
+    UserIdentityType,
+    UserShape
 } from "@/@types/index.ts"
+
+
+export type { Merge, ShapeToObject, EditableShape, InferShape, InferIdentity, InferAuthIdentity } from "@/@types/utility.ts"