+++ packages/core/src/router/context.ts
@@ -1,12 +1,13 @@
 import { createJoseInstance } from "@/jose.ts"
 import { createCookieStore } from "@/cookie.ts"
+import { UserIdentity } from "@/shared/identity.ts"
 import { createProxyLogger } from "@/shared/logger.ts"
 import { createSessionStrategy } from "@/session/strategy.ts"
 import { createBuiltInOAuthProviders } from "@/oauth/index.ts"
 import { getEnv, getEnvArray, getEnvBoolean } from "@/shared/env.ts"
-import type { AuthConfig, InternalContext, User } from "@/@types/index.ts"
+import type { AuthConfig, EditableShape, InternalContext, ShapeToObject, UserShape } from "@/@types/index.ts"
 
-export const createContext = <DefaultUser extends User = User>(config?: AuthConfig): InternalContext<DefaultUser> => {
+export const createContext = <Identity extends EditableShape<UserShape>>(config?: AuthConfig<Identity>) => {
     const trustedProxyHeadersEnv = getEnv("TRUSTED_PROXY_HEADERS")
     const useProxyHeaders =
         trustedProxyHeadersEnv === undefined ? (config?.trustedProxyHeaders ?? false) : getEnvBoolean("TRUSTED_PROXY_HEADERS")
@@ -15,7 +16,7 @@ export const createContext = <DefaultUser extends User = User>(config?: AuthConf
     const cookieOverrides = config?.cookies?.overrides ?? {}
     const secureCookieStore = createCookieStore(true, cookiePrefix, cookieOverrides, logger)
     const standardCookieStore = createCookieStore(false, cookiePrefix, cookieOverrides, logger)
-    const jose = createJoseInstance<DefaultUser>(config?.secret, config?.session)
+    const jose = createJoseInstance<ShapeToObject<Identity>>(config?.secret, config?.session)
 
     const ctx = {
         oauth: createBuiltInOAuthProviders(config?.oauth),
@@ -28,12 +29,18 @@ export const createContext = <DefaultUser extends User = User>(config?: AuthConf
         logger,
         cookieConfig: { secure: secureCookieStore, standard: standardCookieStore },
         baseURL: config?.baseURL,
-    } as InternalContext<DefaultUser>
-    ctx.sessionStrategy = createSessionStrategy({
+        identity: {
+            schema: config?.identity?.schema ?? UserIdentity,
+            unknownKeys: config?.identity?.unknownKeys ?? "strip",
+            skipValidation: config?.identity?.skipValidation ?? false,
+        },
+    } as InternalContext<Identity>
+    ctx.sessionStrategy = createSessionStrategy<Identity>({
         cookies: () => ctx.cookies,
-        jose,
+        jose: ctx.jose,
         config: config?.session,
         logger: ctx.logger,
+        identity: ctx.identity,
     })
     return ctx
 }